refactor(features): refactor redux and lib into feature based hierarchy
