+++ frontend/src/app/codesirius-app.tsx
@@ -4,12 +4,12 @@ import {ReactNode} from 'react'
 import {ThemeProvider} from "@mui/material/styles";
 import {Grid} from "@mui/system";
 import CodesiriusAppBar from "@/components/codesirius-appbar";
-import {createCodeSiriusTheme} from "@/lib/create-theme";
-import {useAppDispatch, useAppSelector} from "@/lib/store/hooks";
 import {AppRouterCacheProvider} from "@mui/material-nextjs/v13-appRouter";
-import {AppDispatch} from "@/lib/store/store";
-import {setUser} from "@/lib/store/codesiriusSlice";
-import {User} from "@/types";
+import {useAppDispatch, useAppSelector} from "@/lib/hooks/hooks";
+import {AppDispatch} from "@/lib/store";
+import {setUser} from "@/lib/features/codesirius/codesiriusSlice";
+import {createCodeSiriusTheme} from "@/lib/utils/create-theme";
+import {User} from "@/lib/features/api/types";
 
 interface CodesiriusThemeProviderProps {
     children: ReactNode;
